+++ addon/gensms.py
@@ -25,6 +25,7 @@ try:
 
     from genmonlib.myconfig import MyConfig
     from genmonlib.mylog import SetupLogger
+    from genmonlib.myclient import ClientInterface
     from genmonlib.mymsgqueue import MyMsgQueue
     from genmonlib.mynotify import GenNotify
     from genmonlib.mysupport import MySupport
@@ -189,6 +190,9 @@ def SendNotice(Message):
 
     try:
 
+        if sitename != None and len(sitename):
+            Message = sitename + ": " + Message
+
         client = Client(account_sid, auth_token)
 
         # send to multiple recipient(s)
@@ -266,6 +270,8 @@ if __name__ == "__main__":
         sys.exit(1)
     try:
 
+        Generator = ClientInterface(host=address, port=port, log=log)
+        sitename = Generator.ProcessMonitorCommand("generator: getsitename")
         Queue = MyMsgQueue(config=config, log=log, callback=SendNotice)
 
         GenNotify = GenNotify(