+++ genmonlib/generac_evolution.py
@@ -4670,7 +4670,7 @@ class Evolution(GeneratorController):
                     }
                 )
 
-            Outage["Outage"].append({"Outage Log": self.DisplayOutageHistory()})
+            Outage["Outage"].append({"Outage Log": self.DisplayOutageHistory(JSONNum=JSONNum)})
 
         except Exception as e1:
             self.LogErrorLine("Error in DisplayOutage: " + str(e1))